+++ packages/smarthr-ui/src/components/Combobox/SingleCombobox/SingleCombobox.tsx
@@ -18,6 +18,7 @@ import {
 import innerText from 'react-innertext'
 import { tv } from 'tailwind-variants'
 
+import { useAnimationFrame } from '../../../hooks/useAnimationFrame'
 import { useClick } from '../../../hooks/useClick'
 import { useLatest } from '../../../hooks/useLatest'
 import { useTheme } from '../../../hooks/useTheme'
@@ -198,7 +199,11 @@ const ActualSingleCombobox = <T,>(
   const [isComposing, setIsComposing] = useState(false)
   const [isEditing, setIsEditing] = useState(false)
 
-  useImperativeHandle<HTMLInputElement | null, HTMLInputElement | null>(ref, () => inputRef.current)
+  useImperativeHandle<HTMLInputElement | null, HTMLInputElement | null>(
+    ref,
+    () => inputRef.current,
+    [],
+  )
 
   const { options } = useSingleOptions({
     items,
@@ -207,6 +212,8 @@ const ActualSingleCombobox = <T,>(
     inputValue,
     isFilteringDisabled: !isEditing,
   })
+
+  const selectFrame = useAnimationFrame()
 
   const { listBoxProps, activeOption, handleKeyDownListBox, listBoxId, listBoxRef } = useListbox<T>(
     {
@@ -220,8 +227,8 @@ const ActualSingleCombobox = <T,>(
         onChangeSelected?.(selected)
 
         // HINT: Dropdown系コンポーネント内でComboboxを使うと、選択肢がportalで表現されている関係上Dropdownが閉じてしまう
-        // requestAnimationFrameを追加、処理を遅延させることで正常に閉じる/閉じないの判定を行えるようにする
-        requestAnimationFrame(() => {
+        // 処理を遅延させることで正常に閉じる/閉じないの判定を行えるようにする
+        selectFrame.request(() => {
           setIsExpanded(false)
           // HINT:
           // - 制御コンポーネントの場合に親側でinputValueを更新できるように、選択時にonChangeInputを空文字で発火する
@@ -240,6 +247,9 @@ const ActualSingleCombobox = <T,>(
       noResultText,
     },
   )
+
+  // TODO: callbackRefにまとめ直したい
+  useEffect(() => selectFrame.cancel, [selectFrame.cancel])
 
   const latest = useLatest({
     onChange,
@@ -393,7 +403,7 @@ const ActualSingleCombobox = <T,>(
       : theme.textColor.grey
 
   useClick(
-    useMemo(() => [triggerRef, listBoxRef, clearButtonRef], [listBoxRef]),
+    [triggerRef, listBoxRef, clearButtonRef],
     isFocused || selectedItem ? NOOP : functions.selectDefaultItem,
     functions.unfocus,
   )