+++ packages/smarthr-ui/src/components/DatePicker/DatePicker.tsx
@@ -18,6 +18,7 @@ import {
 } from 'react'
 import { tv } from 'tailwind-variants'
 
+import { useAnimationFrame } from '../../hooks/useAnimationFrame'
 import { useLatest } from '../../hooks/useLatest'
 import { useOuterClick } from '../../hooks/useOuterClick'
 import { useTheme } from '../../hooks/useTheme'
@@ -148,6 +149,8 @@ export const DatePicker = forwardRef<HTMLInputElement, Props>(
       parseStringDate(value, parseInput),
     )
 
+    const closeFrame = useAnimationFrame()
+
     const latest = useLatest({
       onChange,
       onChangeDate,
@@ -157,6 +160,7 @@ export const DatePicker = forwardRef<HTMLInputElement, Props>(
       onBlur,
       isInputFocused,
       selectedDate,
+      closeFrame,
     })
 
     const functions = useMemo(() => {
@@ -246,7 +250,7 @@ export const DatePicker = forwardRef<HTMLInputElement, Props>(
           if (ESCAPE_KEY_REGEX.test(e.key)) {
             e.stopPropagation()
             // delay hiding calendar because calendar will be displayed when input is focused
-            requestAnimationFrame(closeCalendar)
+            latest.closeFrame.request(closeCalendar)
 
             if (inputRef.current) inputRef.current.focus()
           }
@@ -265,7 +269,7 @@ export const DatePicker = forwardRef<HTMLInputElement, Props>(
         handleSelectDateCalendar: (e: ChangeLikeEvent, selected: Date | null) => {
           updateDate(e, selected)
           // delay hiding calendar because calendar will be displayed when input is focused
-          requestAnimationFrame(closeCalendar)
+          latest.closeFrame.request(closeCalendar)
 
           if (inputRef.current) inputRef.current.focus()
         },
@@ -275,6 +279,7 @@ export const DatePicker = forwardRef<HTMLInputElement, Props>(
     useImperativeHandle<HTMLInputElement | null, HTMLInputElement | null>(
       ref,
       () => inputRef.current,
+      [],
     )
 
     useEffect(() => {
@@ -304,10 +309,7 @@ export const DatePicker = forwardRef<HTMLInputElement, Props>(
       inputRef.current.value = value || ''
     }, [value, isInputFocused, functions])
 
-    useOuterClick(
-      useMemo(() => [inputWrapperRef, calendarPortalRef], [inputWrapperRef, calendarPortalRef]),
-      functions.closeCalendar,
-    )
+    useOuterClick([inputWrapperRef, calendarPortalRef], functions.closeCalendar)
 
     useEffect(() => {
       const handleKeyDown = (e: KeyboardEvent) => {
@@ -358,6 +360,7 @@ export const DatePicker = forwardRef<HTMLInputElement, Props>(
 
       return () => {
         window.removeEventListener('keydown', handleKeyDown)
+        latest.closeFrame.cancel()
       }
     }, [functions, latest])