+++ packages/smarthr-ui/src/hooks/useMediaQueries/useMediaQueries.ts
@@ -1,6 +1,5 @@
-import { useCallback, useMemo, useRef, useSyncExternalStore } from 'react'
+import { useMemo, useRef, useSyncExternalStore } from 'react'
 
-import { entries, fromEntries } from '../../libs/object'
 import { shallowEqual } from '../../libs/shallowEqual'
 
 type MediaQueryListMap = {
@@ -11,54 +10,67 @@ type MediaQueryMatches<T> = {
   [K in keyof T]: boolean
 }
 
+// TODO: EnvironmentProviderからしか利用されていないため、EnvironmentProviderに統合する
 export const useMediaQueries = <T extends MediaQueryListMap>(queries: T): MediaQueryMatches<T> => {
-  const getMatchMediaList = useCallback(
-    () => entries(queries).map(([key, query]) => [key, window.matchMedia(query)] as const),
-    [queries],
-  )
   const lastSnapshotRef = useRef<MediaQueryMatches<T> | null>(null)
 
-  const serverSnapshot = useMemo(
-    () =>
-      fromEntries(entries(queries).map(([key]) => [key, false] as const)) as MediaQueryMatches<T>,
-    [queries],
-  )
+  // useLatest を使わず queries を直接依存配列に指定することで、functions が再作成され、
+  // useSyncExternalStore が subscribe を再実行して新しいメディアクエリを監視できる
+  // queries自体はthemeから取得する想定なので殆どの場合stableだが安全策として依存配列に直接指定する
+  const functions = useMemo(() => {
+    const queryEntries = Object.entries(queries)
+    const serverSnapshot = queryEntries.reduce(
+      (acc, [key]) => {
+        acc[key] = false
+        return acc
+      },
+      {} as Record<string, boolean>,
+    ) as MediaQueryMatches<T>
 
-  const getServerSnapshot = useCallback(
-    () => serverSnapshot,
-    [serverSnapshot],
-  ) satisfies () => MediaQueryMatches<T>
+    return {
+      getServerSnapshot: (() => serverSnapshot) satisfies () => MediaQueryMatches<T>,
+      getSnapshot: (): MediaQueryMatches<T> => {
+        if (typeof window === 'undefined' || !window.matchMedia) {
+          return serverSnapshot
+        }
 
-  const getSnapshot = useCallback((): MediaQueryMatches<T> => {
-    if (typeof window === 'undefined' || !window.matchMedia) {
-      return serverSnapshot
-    }
+        const ret = queryEntries.reduce(
+          (acc, [key, query]) => {
+            acc[key] = window.matchMedia(query).matches
+            return acc
+          },
+          {} as Record<string, boolean>,
+        ) as MediaQueryMatches<T>
 
-    const ret = fromEntries(getMatchMediaList().map(([key, m]) => [key, m.matches] as const))
-    if (lastSnapshotRef.current && shallowEqual(lastSnapshotRef.current, ret)) {
-      return lastSnapshotRef.current
-    }
-    lastSnapshotRef.current = ret
-    return ret
-  }, [getMatchMediaList, serverSnapshot])
+        if (lastSnapshotRef.current && shallowEqual(lastSnapshotRef.current, ret)) {
+          return lastSnapshotRef.current
+        }
 
-  const subscribe = useCallback(
-    (f: () => void) => {
-      if (typeof window === 'undefined' || !window.matchMedia) {
-        return () => {}
-      }
-      const matchMediaList = getMatchMediaList()
-      matchMediaList.forEach(([, m]) => {
-        m.addEventListener('change', f)
-      })
-      return () => {
-        matchMediaList.forEach(([, m]) => {
-          m.removeEventListener('change', f)
+        lastSnapshotRef.current = ret
+
+        return ret
+      },
+      subscribe: (f: () => void) => {
+        if (typeof window === 'undefined' || !window.matchMedia) {
+          return () => {}
+        }
+
+        const mediaQueryList = queryEntries.map(([, query]) => {
+          const m = window.matchMedia(query)
+          m.addEventListener('change', f)
+          return m
         })
-      }
-    },
-    [getMatchMediaList],
-  )
 
-  return useSyncExternalStore(subscribe, getSnapshot, getServerSnapshot)
+        return () => {
+          mediaQueryList.forEach((m) => m.removeEventListener('change', f))
+        }
+      },
+    }
+  }, [queries])
+
+  return useSyncExternalStore(
+    functions.subscribe,
+    functions.getSnapshot,
+    functions.getServerSnapshot,
+  )
 }